+++ frontend/src/systems/AnimationHelper.ts
@@ -1,4 +1,5 @@
 import Phaser from 'phaser';
+import { WALK_DURATION } from '@utils/constants';
 
 /** Helper to register shared sprite animations. */
 export class AnimationHelper {
@@ -14,8 +15,8 @@ export class AnimationHelper {
           start: 0,
           end: 3,
         }),
-        frameRate: 8,
-        repeat: -1,
+        duration: WALK_DURATION,
+        repeat: 0,
       });
 
       scene.anims.create({